requirements.txt
streamlit==1.36.0 
pandas==2.2.2 
plotly==5.15.0 
openpyxl==3.1.5 
Pillow==10.4.0 
xlsxwriter==3.2.0